+++ tui/src/tui/terminal_lib_backends/crossterm_backend/render_op_impl.rs
@@ -32,7 +32,8 @@ use crossterm::{self,
 use r3bl_core::{GCString, InlineVec, LockedOutputDevice, Pos, Size, TuiColor, TuiStyle};
 use smallvec::smallvec;
 
-use crate::{crossterm_color_converter::convert_from_tui_color_to_crossterm_color,
+use r3bl_core::crossterm_color_converter::convert_from_tui_color_to_crossterm_color;
+use crate::{
             disable_raw_mode_now,
             enable_raw_mode_now,
             flush_now,
@@ -307,7 +308,7 @@ mod impl_self {
                 // Handle background color.
                 if let Some(tui_color_bg) = style.color_bg {
                     let color_bg: crossterm::style::Color =
-                        crate::convert_from_tui_color_to_crossterm_color(tui_color_bg);
+                        r3bl_core::convert_from_tui_color_to_crossterm_color(tui_color_bg);
 
                     queue_render_op!(
                         locked_output_device,
@@ -319,7 +320,7 @@ mod impl_self {
                 // Handle foreground color.
                 if let Some(tui_color_fg) = style.color_fg {
                     let color_fg: crossterm::style::Color =
-                        crate::convert_from_tui_color_to_crossterm_color(tui_color_fg);
+                        r3bl_core::convert_from_tui_color_to_crossterm_color(tui_color_fg);
 
                     queue_render_op!(
                         locked_output_device,